skip exporting `TRPCClient`-class + flatten `fetchOpts` (#570)
